feat(devkit): add detectInvokedPackageManager method to devkit
