+++ packages/create-nx-workspace/bin/package-manager.ts
@@ -7,10 +7,24 @@ import { join } from 'path';
  * we duplicate the helper functions from @nrwl/workspace in this file.
  */
 
-const packageManagerList = ['pnpm', 'yarn', 'npm'] as const;
+export const supportedPackageManagers = ['yarn', 'pnpm', 'npm'] as const;
 
-export type PackageManager = typeof packageManagerList[number];
+// Can be changed to "export type PackageManager = typeof supportedPackageManagers[number];" as soon as Typedoc correctly documents it
+export type PackageManager = 'yarn' | 'pnpm' | 'npm';
 
+export interface PackageManagerCommands {
+  install: string;
+  add: string;
+  addDev: string;
+  rm: string;
+  exec: string;
+  list: string;
+  run: (script: string, args: string) => string;
+}
+
+/**
+ * Detects which package manager is used in the workspace based on the lock file.
+ */
 export function detectPackageManager(dir: string = ''): PackageManager {
   return existsSync(join(dir, 'yarn.lock'))
     ? 'yarn'
@@ -29,32 +43,21 @@ export function detectPackageManager(dir: string = ''): PackageManager {
  * ```javascript
  * execSync(`${getPackageManagerCommand().addDev} my-dev-package`);
  * ```
- *
  */
 export function getPackageManagerCommand(
   packageManager: PackageManager = detectPackageManager()
-): {
-  install: string;
-  add: string;
-  addDev: string;
-  rm: string;
-  exec: string;
-  list: string;
-  run: (script: string, args: string) => string;
-} {
-  switch (packageManager) {
-    case 'yarn':
-      return {
-        install: 'yarn',
-        add: 'yarn add -W',
-        addDev: 'yarn add -D -W',
-        rm: 'yarn remove',
-        exec: 'yarn',
-        run: (script: string, args: string) => `yarn ${script} ${args}`,
-        list: 'yarn list',
-      };
-
-    case 'pnpm':
+): PackageManagerCommands {
+  const commands: { [pm in PackageManager]: () => PackageManagerCommands } = {
+    yarn: () => ({
+      install: 'yarn',
+      add: 'yarn add -W',
+      addDev: 'yarn add -D -W',
+      rm: 'yarn remove',
+      exec: 'yarn',
+      run: (script: string, args: string) => `yarn ${script} ${args}`,
+      list: 'yarn list',
+    }),
+    pnpm: () => {
       const [major, minor] = getPackageManagerVersion('pnpm').split('.');
       let useExec = false;
       if (+major >= 6 && +minor >= 13) {
@@ -69,10 +72,10 @@ export function getPackageManagerCommand(
         run: (script: string, args: string) => `pnpm run ${script} -- ${args}`,
         list: 'pnpm ls --depth 100',
       };
+    },
+    npm: () => {
+      process.env.npm_config_legacy_peer_deps ??= 'true';
 
-    case 'npm':
-      process.env.npm_config_legacy_peer_deps =
-        process.env.npm_config_legacy_peer_deps ?? 'true';
       return {
         install: 'npm install',
         add: 'npm install',
@@ -82,40 +85,46 @@ export function getPackageManagerCommand(
         run: (script: string, args: string) => `npm run ${script} -- ${args}`,
         list: 'npm ls',
       };
-  }
+    },
+  };
+
+  return commands[packageManager]();
 }
 
+/**
+ * Returns the version of the package manager used in the workspace.
+ * By default, the package manager is derived based on the lock file,
+ * but it can also be passed in explicitly.
+ */
 export function getPackageManagerVersion(
-  packageManager: PackageManager
+  packageManager: PackageManager = detectPackageManager()
 ): string {
   return execSync(`${packageManager} --version`).toString('utf-8').trim();
 }
 
 /**
- * Detects which package manager was used to invoke create-nx-{plugin|workspace} command
- * based on the main Module process that invokes the command
+ * Detects which package manager was used to invoke the script based on the
+ * main module process that invokes the command.
+ *
  * - npx returns 'npm'
  * - pnpx returns 'pnpm'
  * - yarn create returns 'yarn'
  *
- * Default to 'npm'
+ * By default, 'npm' is returned.
  */
 export function detectInvokedPackageManager(): PackageManager {
   let detectedPackageManager: PackageManager = 'npm';
   // mainModule is deprecated since Node 14, fallback for older versions
   const invoker = require.main || process['mainModule'];
 
-  // default to `npm`
-  if (!invoker) {
-    return detectedPackageManager;
-  }
-
-  for (const pkgManager of packageManagerList) {
-    if (invoker.path.includes(pkgManager)) {
-      detectedPackageManager = pkgManager;
-      break;
+  if (invoker) {
+    for (const pkgManager of supportedPackageManagers) {
+      if (invoker.path.includes(pkgManager)) {
+        return pkgManager;
+      }
     }
   }
 
+  // default to `npm`
   return detectedPackageManager;
 }
